* pretty important bugfix for po8e server:
Basically timesync previously working by grabbing buffer of data and using
the first tick in that buffer and current time to build timesync map.
There were two quirks with this model: 1) should have been using the last
(newest) tick 2) this was only valid when the buffer grab included
all the available data (up until "now").

On the transition to more than 1 po8e card, it became necessary to grab a
fixed k-samples of data on each read (to keep both cards in sync), so there
would be some data "left over" in the po8e buffer. In other words, the wrong
time-tick pair would be used for timesync and on the next read the next
tick would be even more stale. Also it was not clear which was the correct
po8e read to use for sync.

The solution is:
1) only use ticks from the zeroth po8e to sync time.
2) Only grab k-samples, but sync on the newest tick in the po8e buffer,
even if it has not yet been "grabbed"

* Other changes:
- there's no need to send timestamps in zmq buffers; just send ticks
and let clients query the timesync client as needed to get time.
- events are u16 not i16 (otherwise last bit is lost to sign)
- since broadband data is now float32 and is scaled properly,
there's no need to send scaling data